# Backend/openapi_spec.yml
swagger: "2.0"
info:
  title: "simple-script-api"
  description: "API for the backend service of Simple Script"
  version: "1.0.0"
schemes:
  - "https"
produces:
  - application/json
x-google-backend:
  address: https://simple-script-backend-kabidd7ohq-uc.a.run.app
  deadline: 45.0
x-google-endpoints:
- name: https://simple-script-api-bplyx02o.uc.gateway.dev
  allowCors: True
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"
paths:
  /v1/model/gemini-pro:simplify:
    post:
      description: "An endpoint that takes in a string and returns a simplified version of that string by calling Gemini API."
      operationId: "simplify"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "simplifyRequest"
          in: "body"
          description: "Request data containing all necessary parameters"
          required: true
          schema:
            type: "object"
            required:
              - request_id
              - text
              - target_reading_level
            properties:
              request_id:
                type: "string"
              text:
                type: "string"
              target_reading_level:
                type: "string"
      responses:
        200:
          description: "Simplified text returned."
          schema:
            type: "object"
            properties:
              simplifiedText:
                type: "string"
        500:
          description: "Gemini API failed to respond."
      security:
        - api_key: []
    options:
        operationId: "cors"
        responses:
          '200':
            description: A successful response
  /v1/model/gemini-pro:chat:
    post:
      description: "An endpoint that allows a user to chat with Gemini with a webpage!"
      operationId: "chat"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
       - name: "chatRequest"
         in: "body"
         description: "Request data containing all necessary parameters"
         required: true
         schema:
          type: "object"
          required:
            - request_id
            - messages
            - newQuestion
            - webpageContent
          properties:
              request_id:
                type: "string"
              messages:
                type: "object"
              newQuestion:
                type: "string"
              webpageContent:
                type: "string"
      responses:
        200:
          description: "Simplified text returned."
          schema:
            type: "object"
            properties:
              simplifiedText:
                type: "string"
        500:
          description: "Gemini API failed to respond."
      security:
        - api_key: []
      options:
        operationId: "cors-chat"
        responses:
          '200':
            description: A successful response